[url_launcher] Revert embedding support (#2201)
The previous support PR failed to correctly raise the minimum SDK version.
