+++ packages/url_launcher/url_launcher/android/src/main/AndroidManifest.xml
@@ -1,7 +1,7 @@
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
   package="io.flutter.plugins.urllauncher">
   <application>
-    <activity android:name="io.flutter.plugins.urllauncher.WebViewActivity"
+    <activity android:name=".WebViewActivity"
       android:theme="@android:style/Theme.NoTitleBar.Fullscreen"
       android:exported="false"/>
   </application>
