+++ packages/url_launcher/url_launcher/android/src/main/java/io/flutter/plugins/urllauncher/UrlLauncherPlugin.java
@@ -1,82 +1,110 @@
+// Copyright 2017 The Chromium Authors. All rights reserved.
+// Use of this source code is governed by a BSD-style license that can be
+// found in the LICENSE file.
+
 package io.flutter.plugins.urllauncher;
 
-import android.util.Log;
-import androidx.annotation.NonNull;
-import androidx.annotation.Nullable;
-import io.flutter.embedding.engine.plugins.FlutterPlugin;
-import io.flutter.embedding.engine.plugins.activity.ActivityAware;
-import io.flutter.embedding.engine.plugins.activity.ActivityPluginBinding;
+import android.content.ComponentName;
+import android.content.Context;
+import android.content.Intent;
+import android.net.Uri;
+import android.os.Bundle;
+import android.provider.Browser;
+import io.flutter.plugin.common.MethodCall;
+import io.flutter.plugin.common.MethodChannel;
+import io.flutter.plugin.common.MethodChannel.MethodCallHandler;
+import io.flutter.plugin.common.MethodChannel.Result;
 import io.flutter.plugin.common.PluginRegistry.Registrar;
+import java.util.Map;
 
-/**
- * Plugin implementation that uses the new {@code io.flutter.embedding} package.
- *
- * <p>Instantiate this in an add to app scenario to gracefully handle activity and context changes.
- */
-public final class UrlLauncherPlugin implements FlutterPlugin, ActivityAware {
-  private static final String TAG = "UrlLauncherPlugin";
-  private @Nullable MethodCallHandlerImpl methodCallHandler;
-  private @Nullable UrlLauncher urlLauncher;
+/** UrlLauncherPlugin */
+public class UrlLauncherPlugin implements MethodCallHandler {
+  private final Registrar mRegistrar;
 
-  /**
-   * Registers a plugin implementation that uses the stable {@code io.flutter.plugin.common}
-   * package.
-   *
-   * <p>Calling this automatically initializes the plugin. However plugins initialized this way
-   * won't react to changes in activity or context, unlike {@link UrlLauncherPlugin}.
-   */
   public static void registerWith(Registrar registrar) {
-    MethodCallHandlerImpl handler =
-        new MethodCallHandlerImpl(new UrlLauncher(registrar.context(), registrar.activity()));
-    handler.startListening(registrar.messenger());
+    MethodChannel channel =
+        new MethodChannel(registrar.messenger(), "plugins.flutter.io/url_launcher");
+    UrlLauncherPlugin instance = new UrlLauncherPlugin(registrar);
+    channel.setMethodCallHandler(instance);
+  }
+
+  private UrlLauncherPlugin(Registrar registrar) {
+    this.mRegistrar = registrar;
   }
 
   @Override
-  public void onAttachedToEngine(@NonNull FlutterPluginBinding binding) {
-    urlLauncher = new UrlLauncher(binding.getApplicationContext(), /*activity=*/ null);
-    methodCallHandler = new MethodCallHandlerImpl(urlLauncher);
-    methodCallHandler.startListening(binding.getFlutterEngine().getDartExecutor());
+  public void onMethodCall(MethodCall call, Result result) {
+    final String url = call.argument("url");
+    switch (call.method) {
+      case "canLaunch":
+        canLaunch(url, result);
+        break;
+      case "launch":
+        launch(call, result, url);
+        break;
+      case "closeWebView":
+        closeWebView(result);
+        break;
+      default:
+        result.notImplemented();
+        break;
+    }
   }
 
-  @Override
-  public void onDetachedFromEngine(@NonNull FlutterPluginBinding binding) {
-    if (methodCallHandler == null) {
-      Log.wtf(TAG, "Already detached from the engine.");
+  private void canLaunch(String url, Result result) {
+    Intent launchIntent = new Intent(Intent.ACTION_VIEW);
+    launchIntent.setData(Uri.parse(url));
+    ComponentName componentName =
+        launchIntent.resolveActivity(mRegistrar.context().getPackageManager());
+
+    boolean canLaunch =
+        componentName != null
+            && !"{com.android.fallback/com.android.fallback.Fallback}"
+                .equals(componentName.toShortString());
+    result.success(canLaunch);
+  }
+
+  private void launch(MethodCall call, Result result, String url) {
+    Intent launchIntent;
+    final boolean useWebView = call.argument("useWebView");
+    final boolean enableJavaScript = call.argument("enableJavaScript");
+    final boolean enableDomStorage = call.argument("enableDomStorage");
+    final Map<String, String> headersMap = call.argument("headers");
+    final Bundle headersBundle = extractBundle(headersMap);
+    final Context context = mRegistrar.activity();
+
+    if (context == null) {
+      result.error("NO_ACTIVITY", "Launching a URL requires a foreground activity.", null);
       return;
     }
 
-    methodCallHandler.stopListening();
-    methodCallHandler = null;
-    urlLauncher = null;
-  }
-
-  @Override
-  public void onAttachedToActivity(@NonNull ActivityPluginBinding binding) {
-    if (methodCallHandler == null) {
-      Log.wtf(TAG, "urlLauncher was never set.");
-      return;
+    if (useWebView) {
+      launchIntent =
+          WebViewActivity.createIntent(
+              context, url, enableJavaScript, enableDomStorage, headersBundle);
+    } else {
+      launchIntent =
+          new Intent(Intent.ACTION_VIEW)
+              .setData(Uri.parse(url))
+              .putExtra(Browser.EXTRA_HEADERS, headersBundle);
     }
 
-    urlLauncher.setActivity(binding.getActivity());
+    context.startActivity(launchIntent);
+    result.success(true);
   }
 
-  @Override
-  public void onDetachedFromActivity() {
-    if (methodCallHandler == null) {
-      Log.wtf(TAG, "urlLauncher was never set.");
-      return;
+  private void closeWebView(Result result) {
+    Intent intent = new Intent(WebViewActivity.ACTION_CLOSE);
+    mRegistrar.context().sendBroadcast(intent);
+    result.success(null);
+  }
+
+  private Bundle extractBundle(Map<String, String> headersMap) {
+    final Bundle headersBundle = new Bundle();
+    for (String key : headersMap.keySet()) {
+      final String value = headersMap.get(key);
+      headersBundle.putString(key, value);
     }
-
-    urlLauncher.setActivity(null);
-  }
-
-  @Override
-  public void onDetachedFromActivityForConfigChanges() {
-    onDetachedFromActivity();
-  }
-
-  @Override
-  public void onReattachedToActivityForConfigChanges(@NonNull ActivityPluginBinding binding) {
-    onAttachedToActivity(binding);
+    return headersBundle;
   }
 }
