+++ packages/url_launcher/url_launcher/example/android/gradle.properties
@@ -1,4 +1,2 @@
 org.gradle.jvmargs=-Xmx1536M
-android.useAndroidX=true
-android.enableJetifier=true
 android.enableR8=true
