+++ packages/url_launcher/url_launcher/example/pubspec.yaml
@@ -7,13 +7,5 @@ dependencies:
   url_launcher:
     path: ../
 
-dev_dependencies:
-  e2e: "^0.2.0"
-  flutter_driver:
-    sdk: flutter
-
 flutter:
   uses-material-design: true
-
-environment:
-  flutter: ">=1.9.1+hotfix.2 <2.0.0"
